minibank/middleware/requirements.txt
# === Core Python Backend ===
fastapi==0.115.2
uvicorn[standard]==0.30.1

# === Database & ORM ===
SQLAlchemy==2.0.36
pymysql==1.1.1
alembic==1.13.2  # migrasi database MySQL

# === Security / Authentication ===
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0

# === Environment Configuration ===
python-dotenv==1.0.1

# === Tools / Utils / Logging ===
requests==2.32.3
loguru==0.7.2

# === Data Validation ===
pydantic==2.8.2
pydantic-settings==2.4.0

# === Async MySQL Driver (opsional, jika engine async digunakan) ===
aiomysql==0.2.0

# === Testing ===
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2

# === Deployment Option (jika deploy ke produksi) ===
gunicorn==22.0.0
